Add Kafka consumer in Notification Service
This commit introduces the following changes to the codebase:

- Added `NotificationService` class with a method annotated with `@KafkaListener` to consume messages from the `notificationTopic` Kafka topic.
- Added Kafka consumer properties in `application.properties`

Purpose:
The purpose of this commit is to implement a Kafka consumer in the Notification Service application. The Notification Service will consume order notification messages published to the `notificationTopic` Kafka topic by the Order Service. When a message is received, the Notification Service will log the order ID for further processing.